Union Find: Use range.collect() to build vector
Rust includes convenient syntax for working with ranges, as well as
convenient methods for transforming iterators into collections. Instead
of initializing a vec with a fixed set of zeros, we can use a range and
call `collect` on the resulting iterator to build our vector. This
simplifies and shortens the code for the `quick-find` function.