Union Find: Add quick-union algorithm
Adds a naive implementation of the quick-union algorithm. One can
already sense the approach is not an ideal means of expressing this
algorithm in rust. We make several type-casts to ensure we can index
into the vector:

```rust
while i != id[i] as usize {
      i = id[i] as usize;
    }

while j != id[j] as usize {
  j = id[j] as usize;
}
```

There's likely a cleaner way to do this.